// src/interfaces/Article.ts
/*
 * This file is part of the Blog Frontend project.
 * 
 * Copyright (C) 2024 Aniss.dev
 * 
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * 
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * 
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */


// Interface automatically generated by schemas-to-ts

import { Media } from "./Media";
import { Tag } from "./Tag";
import { Media_Plain } from "./Media";
import { Tag_Plain } from "./Tag";
import { Seo, Seo_NoRelations, Seo_Plain } from "./Seo";

export interface Article {
  id: number;
  attributes: {
    createdAt: Date | String;
    updatedAt: Date | String;
    publishedAt?: Date | String;
    title: string;
    coverImage: { data: Media };
    readingTime: string;
    content: string;
    slug: string;
    description: string;
    tags?: { data: Tag[] };
    seo?: Seo;
    locale: string;
    localizations?: { data: Article[] };
  };
}
export interface Article_Plain {
  id: number;
  createdAt: Date | string;
  updatedAt: Date | string;
  publishedAt?: Date | string;
  title: string;
  coverImage: Media_Plain;
  readingTime: string;
  content: string;
  slug: string;
  description: string;
  tags?: Tag_Plain[];
  seo?: Seo_Plain;
  locale: string;
  localizations?: Article_Plain[];
}

export interface Article_NoRelations {
  id: number;
  createdAt: Date | string;
  updatedAt: Date | string;
  publishedAt?: Date | String;
  title: string;
  coverImage: number;
  readingTime: string;
  content: string;
  slug: string;
  description: string;
  tags?: number[];
  seo?: Seo_NoRelations;
  locale: string;
  localizations?: Article[];
}

export interface Article_Preview {
  id: number;
  attributes: {
    publishedAt: string;
    title: string;
    coverImage: { data: Media };
    slug: string;
    description: string;
    tags: { data: Tag[] };
  };
}